Transforms 'execute-api' Resource at Runtime
...Rather Than Retrieving It via Configuration

We can't hang on to the resource directly, as it will cause the cache
to go bad. The user will be authorized only for the single endpoint
and method which their first request specified until the cache entry
expires. How do I know this is how it works? I made that mistake when I
started the danged project!

- Cranks Whips and Chains of Static Analysis Almost to Maximum
- Adds Unit Tests